Updated + Organized all current code
