+++ lib/Models/acconts.dart
@@ -1,0 +1,16 @@
+
+class AccountData {
+  const AccountData({
+    required this.firstName,
+    required this.lastName,
+    required this.phoneNum,
+    required this.email,
+    required this.pass,
+  });
+
+  final String firstName;
+  final String lastName;
+  final int phoneNum;
+  final String email;
+  final String pass;
+}
