--- lib/global_variables.dart
@@ -1,45 +1,0 @@
-final products = [
-  {
-    'id': '0',
-    'eventName': 'Halloween Party',
-    'ticketPrice': 5.00,
-    'imageUrl': 'assets/images/addidas.jpeg',
-    'organisation': 'Nike',
-    'date': '2022-02-01',
-  },
-  {
-    'id': '1',
-    'eventName': 'Tea party',
-    'ticketPrice': 3.12,
-    'imageUrl': 'assets/images/addidas.jpeg',
-    'organisation': 'Addidas',
-    'date': '2025-03-12',
-  },
-  {
-    'id': '2',
-    'eventName': 'placement offers',
-    'ticketPrice': 8.95,
-    'imageUrl': 'assets/images/bataw.jpeg',
-    'organisation': 'Bata',
-    'date': '2023-04-12',
-  },
-  {
-    'id': '3',
-    'eventName': 'Sports fair',
-    'ticketPrice': 4.69,
-    'imageUrl': 'assets/images/jordan.jpeg',
-    'organisation': 'University of Portsmouth',
-    'date': '2022-05-12',
-  },
-];
-
-final cart = [
-  {
-    'id': '3',
-    'eventName': 'Sports fair',
-    'price': 4.69,
-    'imageUrl': 'assets/images/jordan.jpeg',
-    'organisation': 'University of Portsmouth',
-    'date': '2022-05-12',
-  },
-];
