--- lib/home_page.dart
@@ -1,49 +1,0 @@
-import 'package:flutter/material.dart';
-import 'package:uni_pulse/list_events.dart';
-//import 'package:shop_app_flutter/cart_page.dart';
-//import 'package:shop_app_flutter/product_list.dart';
-
-class HomePage extends StatefulWidget {
-  const HomePage({super.key});
-
-  @override
-  State<HomePage> createState() => _HomePageState();
-}
-
-class _HomePageState extends State<HomePage> {
-  int currentPage = 0;
-
-  List<Widget> pages = const [ListEvents()];
-
-  @override
-  Widget build(BuildContext context) {
-    return Scaffold(
-      body: IndexedStack(
-        index: currentPage,
-        children: pages,
-      ),
-      bottomNavigationBar: BottomNavigationBar(
-        iconSize: 35,
-        selectedFontSize: 0,
-        unselectedFontSize: 0,
-        onTap: (value) {
-          setState(() {
-            currentPage = value;
-          });
-        },
-        currentIndex: currentPage,
-        items: const [
-          BottomNavigationBarItem(
-            icon: Icon(Icons.home),
-            label: '',
-          ),
-          BottomNavigationBarItem(
-            icon: Icon(Icons.account_circle),
-            label: 'Profile',
-          ),
-        ],
-      ),
-    );
-  }
-}
-//inherited widget is used for state management by the flutter framework whenever it wants to talk to the parent widget.
