--- lib/register_Screen.dart
@@ -1,218 +1,0 @@
-import 'package:flutter/material.dart';
-// import 'package:firebase_core/firebase_core.dart';
-
-import 'package:uni_pulse_combined/organized/login.dart';
-
-// void main() async {
-//   WidgetsFlutterBinding.ensureInitialized();
-//   try {
-//     await Firebase.initializeApp();
-//   } catch (e) {
-//     print("Firebase Init Error: $e");
-//   }
-//   runApp(const MyApp());
-//  }
-
-// class MyApp extends StatelessWidget {
-//   const MyApp({super.key});
-
-//   @override
-//   Widget build(BuildContext context) {
-//     return MaterialApp(
-//       debugShowCheckedModeBanner: false,
-//       home: HomeScreen(),
-//     );
-//   }
-// }
-
-// class HomeScreen extends StatelessWidget {
-//   const HomeScreen({super.key});
-
-//   @override
-//   Widget build(BuildContext context) {
-//     return Scaffold(
-//       appBar: AppBar(title: const Text("Welcome")),
-//       body: Center(
-//         child: ElevatedButton(
-//           onPressed: () {
-//             print("Navigating to Register Screen..."); // Debugging line
-//             Navigator.push(
-//               context,
-//               MaterialPageRoute(builder: (context) => const RegisterScreen()),
-//             );
-//           },
-//           child: const Text("Go to Register"),
-//         ),
-//       ),
-//     );
-//   }
-// }
-
-class RegisterScreen extends StatefulWidget {
-  const RegisterScreen({super.key});
-
-  @override
-  State<RegisterScreen> createState() {
-    return _RegisterScreenState();
-  }
-}
-
-class _RegisterScreenState extends State<RegisterScreen> {
-  // final _formKey = GlobalKey<FormState>();
-  final TextEditingController _firstNameController = TextEditingController();
-  final TextEditingController _lastNameController = TextEditingController();
-  final TextEditingController _phoneNumberController = TextEditingController();
-  final TextEditingController _emailController = TextEditingController();
-  final TextEditingController _passwordController = TextEditingController();
-  final TextEditingController _confirmPasswordController =
-      TextEditingController();
-  final TextEditingController _dobController = TextEditingController();
-
-  DateTime? _selectedDate;
-
-  Future<void> _selectDate(BuildContext context) async {
-    final DateTime picked = (await showDatePicker(
-      context: context,
-      initialDate: DateTime.now(),
-      firstDate: DateTime(1950),
-      lastDate: DateTime.now(),
-    ))!;
-    if (picked != _selectedDate) {
-      setState(() {
-        _selectedDate = picked;
-        _dobController.text =
-            "${_selectedDate!.day}/${_selectedDate!.month}/${_selectedDate!.year}";
-      });
-    }
-  }
-
-  // Future<void> _registerUser() async {
-  //   if (_formKey.currentState!.validate()) {
-  //     try {
-
-  //       // await FirebaseAuth.instance.createUserWithEmailAndPassword(
-  //       //   email: _emailController.text.trim(),
-  //       //   password: _passwordController.text.trim(),
-  //       // );
-  //       ScaffoldMessenger.of(context).showSnackBar(
-  //         const SnackBar(content: Text("Registration Successful!")),
-  //       );
-  //     } catch (e) {
-  //       ScaffoldMessenger.of(context).showSnackBar(
-  //         SnackBar(content: Text("Error: ${e.toString()}")),
-  //       );
-  //     }
-  //   }
-  // }
-
-  @override
-  Widget build(BuildContext context) {
-    return Scaffold(
-      appBar: AppBar(
-        title: const Text("Register"),
-        backgroundColor: Theme.of(context).colorScheme.inversePrimary,
-      ),
-      body: Padding(
-        padding: const EdgeInsets.all(16.0),
-        child: Form(
-          // key: _formKey,
-          child: Column(
-            children: [
-              TextFormField(
-                controller: _firstNameController,
-                decoration: const InputDecoration(labelText: "First Name"),
-                validator: (value) {
-                  if (value == null || value.isEmpty) {
-                    return "Please enter your first name";
-                  }
-                  return null;
-                },
-              ),
-              TextFormField(
-                controller: _lastNameController,
-                decoration: const InputDecoration(labelText: "Last Name"),
-                validator: (value) {
-                  if (value == null || value.isEmpty) {
-                    return "Please enter your last name";
-                  }
-                  return null;
-                },
-              ),
-              TextFormField(
-                controller: _phoneNumberController,
-                decoration: const InputDecoration(labelText: "Phone Number"),
-                keyboardType: TextInputType.phone,
-                validator: (value) {
-                  if (value == null || value.isEmpty) {
-                    return "Please enter your phone number";
-                  }
-                  if (!RegExp(r'^\+?\d{10,13}$').hasMatch(value)) {
-                    return "Please enter a valid phone number";
-                  }
-                  return null;
-                },
-              ),
-              TextFormField(
-                controller: _emailController,
-                decoration: const InputDecoration(labelText: "Email"),
-                keyboardType: TextInputType.emailAddress,
-                validator: (value) {
-                  if (value == null || value.isEmpty) {
-                    return "Please enter an email";
-                  }
-                  if (!RegExp(r'^[^@]+@[^@]+\.[^@]+').hasMatch(value)) {
-                    return "Please enter a valid email";
-                  }
-                  return null;
-                },
-              ),
-              TextFormField(
-                controller: _passwordController,
-                decoration: const InputDecoration(labelText: "Password"),
-                obscureText: true,
-                validator: (value) {
-                  if (value == null || value.length < 6) {
-                    return "Password must be at least 6 characters";
-                  }
-                  return null;
-                },
-              ),
-              TextFormField(
-                controller: _confirmPasswordController,
-                decoration:
-                    const InputDecoration(labelText: "Confirm Password"),
-                obscureText: true,
-                validator: (value) {
-                  if (value != _passwordController.text) {
-                    return "Passwords do not match";
-                  }
-                  return null;
-                },
-              ),
-              TextFormField(
-                controller: _dobController,
-                decoration: const InputDecoration(labelText: "Date Of Birth"),
-                readOnly: true,
-                onTap: () => _selectDate(context),
-                validator: (value) {
-                  if (_selectedDate == null) {
-                    return "Please select a date of birth";
-                  }
-                  return null;
-                },
-              ),
-              const SizedBox(height: 20),
-              ElevatedButton(
-                onPressed:() {
-                  Navigator.of(context).push(MaterialPageRoute(builder: (ctx) => const AuthScreen() ));}, //_registerUser,
-                child: const Text("Register"),
-              ),
-            ],
-          ),
-        ),
-      ),
-    );
-  }
-}
-
-// test commit
